Resolve env review cleanup issues
