+++ potpie/context-engine/tests/unit/test_build_hook_config.py
@@ -1,12 +1,7 @@
 from __future__ import annotations
 
 import importlib.util
-import os
-import shutil
-import subprocess
 import sys
-import tarfile
-import zipfile
 from pathlib import Path
 from types import ModuleType
 
@@ -60,54 +55,6 @@ def _force_include_source(build_data: dict, destination: str) -> Path:
         if target == destination:
             return Path(source)
     raise AssertionError(f"Missing force_include destination: {destination}")
-
-
-def _archive_text(path: Path, member_suffix: str) -> str:
-    if path.suffix == ".whl":
-        with zipfile.ZipFile(path) as archive:
-            return archive.read(member_suffix).decode("utf-8")
-    if path.name.endswith(".tar.gz"):
-        with tarfile.open(path, "r:gz") as archive:
-            member = next(
-                item
-                for item in archive.getmembers()
-                if item.name.endswith(f"/{member_suffix}")
-            )
-            extracted = archive.extractfile(member)
-            assert extracted is not None
-            return extracted.read().decode("utf-8")
-    raise AssertionError(f"Unsupported build artifact: {path}")
-
-
-def _build_smoke_env() -> dict[str, str]:
-    env = {
-        name: os.environ[name]
-        for name in (
-            "PATH",
-            "HOME",
-            "TMPDIR",
-            "TEMP",
-            "TMP",
-            "UV_CACHE_DIR",
-            "SSL_CERT_FILE",
-            "REQUESTS_CA_BUNDLE",
-        )
-        if name in os.environ
-    }
-    env.update(
-        {
-            "POTPIE_VALIDATE_DISTRIBUTION_DEFAULTS": "1",
-            "POTPIE_ENVIRONMENT": "prod_oss",
-            "POTPIE_SENTRY_DSN": "https://sentry.example.invalid/1",
-            "POTPIE_POSTHOG_API_KEY": "phc_public_smoke",
-            "POTPIE_POSTHOG_HOST": "https://posthog.example.invalid",
-            "LINEAR_CLIENT_ID": "linear-smoke-client",
-            "POTPIE_GITHUB_CLIENT_ID": "github-smoke-client",
-            "POTPIE_BUILD_GIT_SHA": "smoke-sha",
-            "POTPIE_BUILD_TIME": "2026-06-28T00:00:00Z",
-        }
-    )
-    return env
 
 
 def test_distribution_defaults_use_internal_field_names() -> None:
@@ -206,34 +153,6 @@ def test_build_config_environ_overrides_dotenv(tmp_path: Path) -> None:
     assert values["posthog_api_key"] == "env-posthog"
     assert values["linear_client_id"] == "env-linear"
     assert values["github_client_id"] == "env-github"
-
-
-def test_build_config_input_detection(tmp_path: Path) -> None:
-    assert not build_config_values.has_build_config_inputs(
-        build_config_values.DISTRIBUTION_DEFAULT_INPUT_NAMES,
-        {},
-        dotenv_start=tmp_path,
-    )
-
-    assert build_config_values.has_build_config_inputs(
-        build_config_values.DISTRIBUTION_DEFAULT_INPUT_NAMES,
-        {"LINEAR_CLIENT_ID": ""},
-        dotenv_start=tmp_path,
-    )
-
-    (tmp_path / ".env").write_text("UNRELATED=1\n", encoding="utf-8")
-    assert not build_config_values.has_build_config_inputs(
-        build_config_values.DISTRIBUTION_DEFAULT_INPUT_NAMES,
-        {},
-        dotenv_start=tmp_path,
-    )
-
-    (tmp_path / ".env").write_text("POTPIE_SENTRY_DSN=file-sentry\n", encoding="utf-8")
-    assert build_config_values.has_build_config_inputs(
-        build_config_values.DISTRIBUTION_DEFAULT_INPUT_NAMES,
-        {},
-        dotenv_start=tmp_path,
-    )
 
 
 def test_distribution_defaults_have_safe_public_defaults() -> None:
@@ -440,63 +359,6 @@ def test_distribution_defaults_hook_refuses_unexpected_cleanup_path(
             {"_potpie_generated_build_dirs": [str(tmp_path / "not-owned")]},
             str(tmp_path / "dist.whl"),
         )
-
-
-def test_distribution_defaults_hook_builds_wheel_and_sdist_with_generated_modules(
-    tmp_path: Path,
-) -> None:
-    uv = shutil.which("uv")
-    if uv is None:
-        pytest.skip("uv is required for the packaging smoke test")
-    context_engine = Path(__file__).resolve().parents[2]
-
-    result = subprocess.run(
-        [uv, "build", "--out-dir", str(tmp_path)],
-        cwd=context_engine,
-        env=_build_smoke_env(),
-        text=True,
-        capture_output=True,
-        check=False,
-    )
-
-    assert result.returncode == 0, result.stderr
-    artifacts = list(tmp_path.iterdir())
-    wheel = next(path for path in artifacts if path.suffix == ".whl")
-    sdist = next(path for path in artifacts if path.name.endswith(".tar.gz"))
-    for artifact in (wheel, sdist):
-        distribution_defaults = _archive_text(
-            artifact, "bootstrap/_distribution_defaults.py"
-        )
-        build_info = _archive_text(artifact, "bootstrap/_build_info.py")
-        assert "DISTRIBUTION_DEFAULTS = {" in distribution_defaults
-        assert "'environment': 'prod_oss'" in distribution_defaults
-        assert "'sentry_dsn': 'https://sentry.example.invalid/1'" in (
-            distribution_defaults
-        )
-        assert "'posthog_api_key': 'phc_public_smoke'" in distribution_defaults
-        assert "'posthog_host': 'https://posthog.example.invalid'" in (
-            distribution_defaults
-        )
-        assert "'linear_client_id': 'linear-smoke-client'" in distribution_defaults
-        assert "'github_client_id': 'github-smoke-client'" in distribution_defaults
-        assert "GIT_SHA = 'smoke-sha'" in build_info
-        assert "BUILD_TIME = '2026-06-28T00:00:00Z'" in build_info
-    assert not (context_engine / "bootstrap" / "_distribution_defaults.py").exists()
-    assert not (context_engine / "bootstrap" / "_build_info.py").exists()
-
-
-def test_prefer_existing_build_info(tmp_path: Path) -> None:
-    out = tmp_path / "_build_info.py"
-    build_config_values.write_python_constants(
-        out, {"GIT_SHA": "old", "BUILD_TIME": "kept"}
-    )
-
-    values = build_config_values.prefer_existing_config_values(
-        out,
-        {"GIT_SHA": "", "BUILD_TIME": "default", "EXTRA": "new"},
-    )
-
-    assert values == {"GIT_SHA": "old", "BUILD_TIME": "kept", "EXTRA": "new"}
 
 
 def test_prefer_existing_build_info_preserves_missing_field_inputs(
