+++ potpie/context-engine/tests/unit/test_env_bootstrap.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from adapters.outbound.cli_auth import env_bootstrap as eb
+from bootstrap import env_bootstrap as eb
 
 
 def test_parse_env_line() -> None:
